Put the getmatchword test cases into Autocompletion.test.ts
